+++ 3D-Graph-Learn/GraphUtils.h
@@ -8,8 +8,8 @@
 // TODO: Maybe it is a good idea to make this file main node of all graph files
 namespace vgu {
 	template <unsigned char N>
-	float dotProduct(const Vector<N>& A, const Vector<N>& B) {
-		float sum = 0;
+	double dotProduct(const Vector<N>& A, const Vector<N>& B) {
+		double sum = 0;
 		for (int i = 0; i < N; ++i) sum += A[i] * B[i];
 		return sum;
 	}
@@ -19,22 +19,22 @@ namespace vgu {
 		return A / A.length();
 	}
 
-	float calcDetRecursively(const float values[], const int& size);
+	double calcDetRecursively(const double values[], const int& size);
 
 	template <unsigned char N>
-	float det(const Matrix<N, N>& mat) {
+	double det(const Matrix<N, N>& mat) {
 		return calcDetRecursively(mat.m_values, N);
 	}
 
 	template <unsigned char N = 2>
-	float det(const Matrix<2, 2>& mat) {
+	double det(const Matrix<2, 2>& mat) {
 		return mat(0, 0) * mat(1, 1) - mat(0, 1) * mat(1, 0);
 	}
 
 	template <unsigned char N = 3>
-	float det(const Matrix<3, 3>& mat) {
-		const float firstSum = mat(0, 0) * mat(1, 1) * mat(2, 2) + mat(0, 1) * mat(1, 2) * mat(2, 0) + mat(1, 0) * mat(2, 1) * mat(0, 2);
-		const float secondSum = mat(0, 2) * mat(1, 1) * mat(2, 0) + mat(1, 0) * mat(0, 1) * mat(2, 2) + mat(2, 1) * mat(1, 2) * mat(0, 0);
+	double det(const Matrix<3, 3>& mat) {
+		const double firstSum = mat(0, 0) * mat(1, 1) * mat(2, 2) + mat(0, 1) * mat(1, 2) * mat(2, 0) + mat(1, 0) * mat(2, 1) * mat(0, 2);
+		const double secondSum = mat(0, 2) * mat(1, 1) * mat(2, 0) + mat(1, 0) * mat(0, 1) * mat(2, 2) + mat(2, 1) * mat(1, 2) * mat(0, 0);
 		return firstSum - secondSum;
 	}
