+++ 3D-Graph-Learn/Matrix.h
@@ -16,10 +16,10 @@ namespace vgu {
 	Matrix<M, N>& operator-(Matrix<M, N> A, const Matrix<M, N>& B);
 
 	template <unsigned int M, unsigned int N>
-	Matrix<M, N>& operator*(Matrix<M, N> mat, const float& value);
+	Matrix<M, N>& operator*(Matrix<M, N> mat, const double& value);
 
 	template <unsigned int M, unsigned int N>
-	Matrix<M, N>& operator*(const float& value, Matrix<M, N> mat);
+	Matrix<M, N>& operator*(const double& value, Matrix<M, N> mat);
 
 	template <unsigned int M, unsigned int N>
 	Vector<N> operator*(const Vector<M>& vec, const Matrix<M, N>& mat);
@@ -31,47 +31,47 @@ namespace vgu {
 	template <unsigned int M, unsigned int N>
 	class Matrix {
 	public:
-		float m_values[M * N];
+		double m_values[M * N];
 
 		Matrix();
-		Matrix(const float& value);
+		Matrix(const double& value);
 		~Matrix() = default;
 
 		// TODO: Find a way to move this constructor to template file
 		template <typename ...T>
 		Matrix(T... data) {
 			const unsigned int size = M * N;
-			const float arr[size]{ static_cast<float>(data)... };
+			const double arr[size]{ static_cast<double>(data)... };
 			for (int i = 0; i < size; ++i) {
 				m_values[i / N + M * (i % N)] = arr[i];
 			}
 		}
 
 		Matrix<M, N>& operator=(const Matrix<M, N>& mat);
-		float& operator[](const int& index);
-		float& operator[](int&& index);
-		const float& operator[](const int& index) const;
-		const float& operator[](int&& index) const;
-		float& operator()(const int& mIndex, const int& nIndex);
-		float& operator()(int&& mIndex, int&& nIndex);
-		const float& operator()(const int& mIndex, const int& nIndex) const;
-		const float& operator()(int&& mIndex, int&& nIndex) const;
+		double& operator[](const int& index);
+		double& operator[](int&& index);
+		const double& operator[](const int& index) const;
+		const double& operator[](int&& index) const;
+		double& operator()(const int& mIndex, const int& nIndex);
+		double& operator()(int&& mIndex, int&& nIndex);
+		const double& operator()(const int& mIndex, const int& nIndex) const;
+		const double& operator()(int&& mIndex, int&& nIndex) const;
 		friend Matrix<M, N>& operator+ <> (Matrix<M, N> A, const Matrix<M, N>& B);
 		friend Matrix<M, N>& operator- <>(Matrix<M, N> A, const Matrix<M, N>& B);
-		friend Matrix<M, N>& operator* <>(Matrix<M, N> mat, const float& value);
-		friend Matrix<M, N>& operator* <>(const float& value, Matrix<M, N> mat);
+		friend Matrix<M, N>& operator* <>(Matrix<M, N> mat, const double& value);
+		friend Matrix<M, N>& operator* <>(const double& value, Matrix<M, N> mat);
 		friend Vector<N> operator* <>(const Vector<M>& vec, const Matrix<M, N>& mat);
 		friend Vector<M> operator* <>(const Matrix<M, N>& mat, const Vector<N>& vec);
 		Matrix<N, N> operator-() const;
 		Matrix<M, N>& operator+=(const Matrix<M, N>& mat);
 		Matrix<M, N>& operator-=(const Matrix<M, N>& mat);
-		Matrix<M, N>& operator*=(const float& value);
-		Matrix<M, N>& operator/=(const float& value);
+		Matrix<M, N>& operator*=(const double& value);
+		Matrix<M, N>& operator/=(const double& value);
 
 		template <unsigned int U>
 		Matrix<M, U>& operator*=(const Matrix<N, U>& matrix) {
 			const unsigned int size = M * N;
-			float tempValues[size];
+			double tempValues[size];
 			for (int i = 0; i < size; ++i) {
 				tempValues[i] = m_values[i];
 				m_values[i] = 0;
